Solve day 5
This is just a reinforcement that the right data structure makes all
the difference. I thought of implementing this in terms of vectors,
but that would be quite a lot of copying. Linked lists, though they
are usually not the right choice, were in fact the right choice today.

$ hyperfine --warmup 5 'target/release/day05'
Wed Dec  5 13:11:26 STD 2018
Benchmark #1: target/release/day05
  Time (mean ± σ):     109.1 ms ±   1.5 ms    [User: 96.2 ms, System: 10.5 ms]
  Range (min … max):   107.6 ms … 113.5 ms