first-pass optimize day 9
Yeah, allocations were a substantial part of the runtime.
Still not sure where the rest of it comes from, but I suppose we'd need
proper profiling to answer that question. I might come back to it later
armed with a profiler to find out... but I probably won't.

$ hyperfine --warmup 5 target/release/day09
Sun Dec  9 17:01:18 STD 2018
Benchmark #1: target/release/day09
  Time (mean ± σ):     615.6 ms ±   8.4 ms    [User: 252.9 ms, System: 344.9 ms]
  Range (min … max):   594.2 ms … 627.7 ms

It does feel pretty good to have cut 40% off the runtime with a half-dozen
lines of code.