second-pass optimize day 9
I keep reaching for fancy list-based data structures when what I really need,
apparently, is VecDeque. The VecDeque implementation is not only substantially
shorter than the doubly-linked list implementation, it cuts the runtime by 2/3.

This is still a lot slower than it might be--I keep hearing about implementations
finishing the exercise in 60 milliseconds or so--but this is good enough for me:
less than twice as slow as the next slowest exercise.

For now.

$ hyperfine --warmup 5 target/release/day09
Mon Dec 10 22:58:42 STD 2018
Benchmark #1: target/release/day09
  Time (mean ± σ):     194.7 ms ±   5.1 ms    [User: 108.0 ms, System: 71.1 ms]
  Range (min … max):   189.0 ms … 205.5 ms