start rewriting day12 in modern style
This one is a pretty substantial one, it seems; each iteration
should be pretty straightforward, but automatically discovering the
pattern and generalizing to 50 billion generations is where the
challenge is.

Also, parsing the input is kind of a pain, so we'll have to see how that goes.
Stole the fundamental parser from the old version, but it looks like
I'd kind of cheated there in the past; may need to get fancy.