solve day13 part 2
Turns out that when the elves work instantaneously, they mean
_instantaneously_: if two carts collide with a third incoming
on the same tick, the two are gone and the third passes through
without obstruction.

That meant that I had to rewrite the tick architecture a bit to
appease the borrow-checker, but it does appear to be both
sufficient and efficient.