solve day15 part 2
I wasn't 100% confident, when I was first laying out the problem,
that it was sensible to give each unit its own attack power. On
the other hand, it was cheap, and I could always take it out later.
Turns out to have been a bit prescient; part 2 was really easy,
and succeeded on the first try.

This whole problem could use another pass; there are definitely
places where it could be more efficient. However, I'm done, for
now at least. Despite, or even because of, the thorough specification,
this one felt kind of tediously finicky.