WIP: implement water propagation algorithm
Unfortunately, there's an architectural flaw with this version:
when we attack the problem purely procedurally, it's very hard to
backtrack when we discover that it's time to fill a level of a basin.

We need some kind of recursive implementation which returns a status
enum: { Terminated, Fill }. If filling outwards from any level causes
a basin to fill, then the caller needs to fill itself outwards also.